detection/requirements.txt
Pillow>=12.1.0
numpy>=2.4.2
onnxruntime>=1.23.2
PyYAML>=6.0.3
